refactor(ipc): split contract types from transport runtime
Extract all IPC type/interface declarations into a dedicated
ipc-contract.ts file, keeping only transport/runtime helpers
(serialize, createMessageParser, MAX_LINE_SIZE) in ipc-protocol.ts.

- Remove duplicate MessageQueued/MessageDequeued declarations
- Add IPCContractSchema export for future codegen consumption
- Re-export all contract types from ipc-protocol.ts for backward compat
- Add compile-time type compatibility smoke test

Closes #1997
